requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cssselect==1.2.0
exceptiongroup==1.1.3
fake-useragent==1.3.0
frozenlist==1.4.0
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
joblib==1.3.2
lxml==4.9.3
multidict==6.0.4
nltk==3.8.1
oauthlib==3.2.2
openai==0.28.1
outcome==1.3.0.post0
parse==1.19.1
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyee==8.2.2
pygsheets==2.0.6
pyparsing==3.1.1
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
uritemplate==4.1.1
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
